pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "alakazam"
version = "1.0.0"
description = "ALAKAZAM: Radio interferometry calibration with chain solver"
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.20",
    "scipy>=1.7",
    "numba>=0.55",
    "h5py>=3.0",
    "pyyaml>=5.0",
    "python-casacore>=3.4",
    "rich>=10.0",
]

[project.scripts]
alakazam = "alakazam.cli:main"

[tool.setuptools.packages.find]
include = ["alakazam*"]
